pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "personalfinance"
version = "0.1.0"
description = "Privacy-first personal finance toolkit built on Beancount v3, with MCP server and CLI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Ankur Aggarwal" },
]
keywords = ["beancount", "finance", "accounting", "mcp", "double-entry"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business :: Financial :: Accounting",
]
dependencies = [
    "beancount>=3.0.0",
    "beanquery>=0.2",
    "beangulp>=0.2",
    "mcp[cli]>=1.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pdfplumber>=0.11",
    "openpyxl>=3.1",
    "xlrd>=2.0",
    "python-dotenv>=1.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
market = [
    "yfinance>=0.2",
    "pandas>=2.0",
]
agent = [
    "rich>=13.0",
    "ollama>=0.3",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
finkit = "personalfinance.cli:main"

[project.urls]
Homepage = "https://github.com/ankuragwl/finkit"
Repository = "https://github.com/ankuragwl/finkit"

[tool.hatch.build.targets.wheel]
packages = ["src/personalfinance"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
